Remove Delaunay2DGridFit
This class was never really used as Delaunay is a natural scatter
fitter. It was only there because the nn_interpolator option in the
old delaunay module had two modes: interp(x0, x1) for unstructured
data and interp[x0, x1] for gridded data. Both could be handled by
the former, but this had issues in the old module, resulting in the
special case.