Depend on scipy >= 0.9.0
This ensures that the Delaunay interpolators are available in
scipy.interpolate. In addition, remove all checks for scipy 0.7.0
involving PiecewisePolynomial and Rbf interpolators as this will
automatically be satisfied. Coincidentally this picked up a bug
where PiecewisePolynomial1DFit wasn't tested anymore as the old
underlying class disappeared from the latest scipy.

The dependency is not particularly onerous as scipy 0.9.0 dates
back to 2011... Also, there is no need for matplotlib anymore.