SPR1-2994 Strip out numpy.distutils
This is lost cargo from the Cult of Scikits-Image. I can't figure out
what it's used for (skimage doesn't even have a setup.py anymore!).

It is a pain though because distutils is deprecated and slated to be
removed in Python 3.12 (see PEP 632). Even worse, scikits.fitting uses
NumPy's version in setup.py which makes NumPy a setup requirement.
This in turn makes install_pinned.py unhappy and blocks katsdpcalproc.