SPR1-2994 Move to pkgutil-style namespace package
We previously used the old pkg_resources style but the setup() kwarg
`namespace_packages` pops up warnings in setuptools now. A native
namespace package aka PEP 420 is also appealing but it only appeared in
Python 3.3 and we still support Python 2.7... The pkgutil style is
therefore a good compromise between compatibility and modernity.

For more info on the choices, see the discussion at

https://packaging.python.org/en/latest/guides/packaging-namespace-packages/

There will be an issue co-existing with other ancient scikits packages
that still use pkg_resources but I couldn't find any to check this.

Many (most?) scikits packages dropped the namespace and contracted the
name to "skxxxxx" (e.g. sklearn, skimage). This suggests that we rename
the package to "skfitting" but given its limited use I prefer to forge
ahead with what we have.

Also add an __init__.py to the tests directory to make it automatically
discoverable as a package.